Changes to writing/reading algorithm
A block size is now considered when writing/reading. This block size
is used to work with compression using the h.264 codec. The DCT algo
assigns mean values to blocks of data. By setting the data to the
same value over a block, the mean is generated to find out what data
was inserted.

Further improvements have been made such as moving the read/write
y position if the width of the frame has been reached.

Lodge header has been shortened to minimise the data taken up by
it, as we prepare for additional data to be stored inside.

Tests were updated to make them work with the new flow.

A mess of log lines will be cleaned up in the next commit.