Frame header now is set on each frame
Frame headers are generated for each frame.
The header contains data about that current
frame and the steg data inside it. The tot-
al amount of steg data is stored, along
with the frame position, total frames,and
subtitle name.

In the next commit, block size might be kept
in the frame header as well.

The total amount of steg data inside a
frame is worked out by converting a temp
frame header to string and getting it's
size. Then subtracting the total data
allowed inside a frame with the frame
header size.